reuse block file on db upgrade

+++ src/txdb-leveldb.cpp
@@ -34,13 +34,14 @@ static leveldb::Options GetOptions() {
     return options;
 }
 
-void init_blockindex(leveldb::Options& options, bool fRemoveOld = false) {
+static void init_blockindex(leveldb::Options& options, bool fRemoveOld = false, bool fCreateBootstrap = false) {
     // First time init.
     filesystem::path directory = GetDataDir() / "txleveldb";
 
     if (fRemoveOld) {
         filesystem::remove_all(directory); // remove directory
         unsigned int nFile = 1;
+		filesystem::path bootstrap = GetDataDir() / "bootstrap.dat";
 
         while (true)
         {
@@ -50,7 +51,11 @@ void init_blockindex(leveldb::Options& options, bool fRemoveOld = false) {
             if( !filesystem::exists( strBlockFile ) )
                 break;
 
-            filesystem::remove(strBlockFile);
+            if (fCreateBootstrap && nFile == 1 && !filesystem::exists(bootstrap)) {
+                filesystem::rename(strBlockFile, bootstrap);
+            } else {
+                filesystem::remove(strBlockFile);
+            }
 
             nFile++;
         }
@@ -101,7 +106,7 @@ CTxDB::CTxDB(const char* pszMode)
             delete activeBatch;
             activeBatch = NULL;
 
-            init_blockindex(options, true); // Remove directory and create new database
+            init_blockindex(options, true, true); // Remove directory and create new database
             pdb = txdb;
 
             bool fTmp = fReadOnly;
